+++ lib/features/wp-rets-client/lib/classes/class-xmlrpc.php
@@ -470,9 +470,9 @@ namespace UsabilityDynamics\WPRETSC {
 
         foreach( (array) $_post_data_tax_input as $tax_name => $tax_tags){
 
-          // Ignore these taxonomies if we support [wpp_location].
-          if( defined( 'WPP_FEATURE_FLAG_WPP_LOCATION' ) && WPP_FEATURE_FLAG_WPP_LOCATION && in_array( $tax_name, array( 'rets_location_state', 'rets_location_county', 'rets_location_city', 'rets_location_route' ) ) ) {
-            ud_get_wp_rets_client()->write_log( "Skipping [$tax_name] taxonomy, we have [wpp_location] enabled." );
+          // Ignore these taxonomies if we support [wpp_listing_location].
+          if( defined( 'WPP_FEATURE_FLAG_WPP_LISTING_LOCATION' ) && WPP_FEATURE_FLAG_WPP_LISTING_LOCATION && in_array( $tax_name, array( 'rets_location_state', 'rets_location_county', 'rets_location_city', 'rets_location_route' ) ) ) {
+            ud_get_wp_rets_client()->write_log( "Skipping [$tax_name] taxonomy, we have [wpp_listing_location] enabled." );
             continue;
           }
 
@@ -489,9 +489,9 @@ namespace UsabilityDynamics\WPRETSC {
 
           }
 
-          // If WP-Property location flag is enabled, and we're doing the [wpp_location] taxonomy, and the WPP_F::update_location_terms method is callable, process our wpp_location terms.
-          if( defined( 'WPP_FEATURE_FLAG_WPP_LOCATION' ) && WPP_FEATURE_FLAG_WPP_LOCATION && $tax_name === 'wpp_location' && is_callable(array( 'WPP_F', 'update_location_terms' ) ) ) {
-            ud_get_wp_rets_client()->write_log( 'Handling [wpp_location] taxonomy for [' . $_post_id .'] listing.' );
+          // If WP-Property location flag is enabled, and we're doing the [wpp_listing_location] taxonomy, and the WPP_F::update_location_terms method is callable, process our wpp_listing_location terms.
+          if( defined( 'WPP_FEATURE_FLAG_WPP_LISTING_LOCATION' ) && WPP_FEATURE_FLAG_WPP_LISTING_LOCATION && $tax_name === 'wpp_listing_location' && is_callable(array( 'WPP_F', 'update_location_terms' ) ) ) {
+            ud_get_wp_rets_client()->write_log( 'Handling [wpp_listing_location] taxonomy for [' . $_post_id .'] listing.' );
 
             $_geo_tag_fields = array(
               "state" => isset( $_post_data_tax_input["rets_location_state"] ) ? reset( $_post_data_tax_input["rets_location_state"] ) : null,
@@ -509,7 +509,7 @@ namespace UsabilityDynamics\WPRETSC {
               ud_get_wp_rets_client()->write_log( "Inserted " . count( $_location_terms ) . " location terms for [" .$_post_id."] property." );
             }
 
-            // Avoid re-adding whatever fields were passed by real [wpp_location]
+            // Avoid re-adding whatever fields were passed by real [wpp_listing_location]
             continue;
 
           }